src/requirements.txt
prometheus-client==0.19.0
pyTenable==1.4.18